Add better support for angle conversions
Adds the `degrees` and `radians` postfix factor modifiers.

The calculator now tries to guess whether-or-not you want to see
the angle in the output (i.e, `rad` after the numeric).

Bit hacky, and I'm sure there'll be bugs with this. Hopefully this will
be made more robust when I jazz up the result package.